# cedar-monitor-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nThe monitor server reports on the rest of CEDAR: how deep the work queues are, how much each\nstore holds, whether a given server considers itself healthy, and what artifact a given string\nrefers to.\n\nEverything here is read-only and needs the monitor read permission. The counts are meant to be\nread against each other: the workspace graph is authoritative, and the search index disagreeing\nwith it is how a half-finished reindex shows itself.\n",
    "title" : "CEDAR Monitor Server API",
    "version" : "2.9.10-SNAPSHOT"
  },
  "paths" : {
...
    "/logs/usage/summary" : {
      "get" : {
        "description" : "The totals over the range beside the hourly series they were summed from. Both come from the hourly rollups, so the newest hour appears only once the aggregator has closed it. Requires the monitor read permission.",
        "operationId" : "summary",
        "parameters" : [ {
          "description" : "Inclusive lower bound, as an ISO-8601 instant in UTC. Defaults to seven days before the upper bound.",
          "in" : "query",
          "name" : "from",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Exclusive upper bound, as an ISO-8601 instant in UTC. Defaults to now.",
          "in" : "query",
          "name" : "to",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LogUsageSummary"
                }
              }
            },
            "description" : "Totals and the hourly series behind them"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The log database could not be read, or a bound was not an ISO-8601 instant"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get request volume and latency totals for a time range",
        "tags" : [ "Log usage" ]
      }
    },
...
    "/resource/users" : {
      "get" : {
        "description" : "Gather what each store holds about one user into a single answer: the workspace graph's record of it, the counts it can reach in each store, its group memberships, and the Keycloak account behind it. Written for diagnosis rather than for an application: the point is to see the stores side by side, since a user that behaves oddly usually has one store disagreeing with another. A store that cannot be reached leaves its section null rather than failing the request, and an identifier nothing knows returns an empty answer with 200.",
        "operationId" : "getUserInfo",
        "parameters" : [ {
          "description" : "Identifier of the user to report on.",
          "in" : "query",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDiagnosticReport"
                }
              }
            },
            "description" : "What each store holds about the user"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Get everything CEDAR knows about a user",
        "tags" : [ "Diagnostics" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CedarError" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "A CEDAR error response. Diagnostic fields are populated only when they are relevant to the failure.",
        "properties" : {
          "entities" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named domain entities associated with exception-mapped errors.",
              "nullable" : true
            },
            "description" : "Named domain entities associated with exception-mapped errors.",
            "nullable" : true
          },
          "errorId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "Correlation identifier for an internal error recorded in server logs.",
            "nullable" : true
          },
          "errorKey" : {
            "type" : "string",
            "description" : "Stable machine-readable error identifier.",
            "enum" : [ "templateElementNotCreated", "templateElementNotFound", "templateElementNotDeleted", "templateElementNotUpdated", "templateElementsNotListed", "templateFieldNotCreated", "templateFieldNotFound", "templateFieldNotDeleted", "templateFieldNotUpdated", "templateFieldsNotListed", "templateNotCreated", "templateNotFound", "templateNotDeleted", "templateNotUpdated", "templatesNotListed", "templateInstanceNotCreated", "templateInstanceNotFound", "templateInstanceNotDeleted", "templateInstanceNotUpdated", "templateInstancesNotListed", "noReadAccessToResource", "noWriteAccessToResource", "noReadAccessToFolder", "noWriteAccessToFolder", "noReadAccessToTemplate", "noWriteAccessToTemplate", "noReadAccessToTemplateElement", "noWriteAccessToTemplateElement", "noReadAccessToTemplateField", "noWriteAccessToTemplateField", "noReadAccessToTemplateInstance", "noWriteAccessToTemplateInstance", "noReadAccessToArtifact", "noWriteAccessToArtifact", "noReadAccessToCategory", "noWriteAccessToCategory", "folderNotFound", "artifactNotFound", "artifactPreconditionRequired", "artifactHasMovedOn", "verbatimWriteRefused", "nodeNotFound", "unknownResourceType", "missingParameter", "invalidInput", "sourceFolderNotFound", "sourceResourceNotFound", "targetFolderNotFound", "nodeNotMoved", "groupNotFound", "userNotFound", "uniqueConstraintCollision", "invalidData", "notAuthorized", "permissionMissing", "tokenInvalid", "userInfoLoadByTokenFailed", "userInfoLoadByApiKeyFailed", "cedarUserNotFound", "authorizationNotFound", "permissionNotOwned", "authorizationTypeUnknown", "apiKeyNotFound", "tokenMissing", "tokenExpired", "parentFolderNotSpecified", "parentFolderSpecifiedTwice", "pathNotNormalized", "parentFolderNotFound", "updateInvalidFolderName", "createInvalidFolderName", "nodeAlreadyPresent", "folderNotCreated", "missingNameAndDescription", "missingData", "folderNotDeleted", "artifactNotDeleted", "resourceNotCreated", "invalidResourceType", "invalidArtifactType", "readOtherProfileForbidden", "updateOtherProfileForbidden", "folderCanNotBeDeleted", "folderCanNotBeChanged", "groupAlreadyPresent", "groupCanBeModifiedOnlyByGroupAdmin", "groupCanBeDeletedOnlyByGroupAdmin", "groupMembersCanBeReadOnlyByGroupAdmin", "groupRequiresAdministrator", "groupUsersNotUpdated", "specialGroupCanNotBeDeleted", "folderPermissionsCanNotBeChanged", "unknownInstanceOutputFormat", "folderCopyNotAllowed", "methodNotImplemented", "upstreamServerError", "nothingToDo", "parentCategoryNotFound", "categoryAlreadyPresent", "categoryNotFound", "categoryCanNotBeDeleted", "rootCategoryCanNotBeDeleted", "unableToAttachCategory", "unableToDetachCategory", "noCategoriesWereAttached", "malformedJsonRequestBody", "malformedSearchTerm", "publishedArtifactCanNotBeChanged", "publishedArtifactCanNotBeDeleted", "versioningOnlyOnLatest", "versioningOnlyByOwner", "nonVersionedArtifactType", "createDraftOnlyFromPublished", "publishOnlyDraft", "draftNotCreated", "contentNotValid", "doiNotSupportedByResourceType", "doiCanNotBeSetForEmptyAtId", "doiCanNotBeAltered", "doiCanNotBeSet", "doiAlreadyExists", "dataCiteDOIDisabled", "resourceNotFound", "internalError" ],
            "nullable" : true
          },
          "errorMessage" : {
            "type" : "string",
            "description" : "Alias of message retained for compatibility.",
            "nullable" : true
          },
          "errorReasonKey" : {
            "type" : "string",
            "description" : "More specific machine-readable reason for the error.",
            "enum" : [ "nonEmptyFolder", "nonEmptyCategory", "userHomeFolder", "systemFolder", "templateReferencedInInstances", "validationError", "continuationExpired" ],
            "nullable" : true
          },
          "errorType" : {
            "type" : "string",
            "description" : "Broad category of the error.",
            "enum" : [ "notFound", "invalidArgument", "authorization", "permission", "server", "validationError" ],
            "nullable" : true
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message.",
            "nullable" : true
          },
          "objects" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named structured values associated with the error."
            },
            "description" : "Named structured values associated with the error."
          },
          "operation" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Operation that failed. Its fields depend on the operation type.",
              "nullable" : true
            },
            "description" : "Operation that failed. Its fields depend on the operation type.",
            "nullable" : true
          },
          "parameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "description" : "Named scalar values associated with the error."
            },
            "description" : "Named scalar values associated with the error."
          },
          "status" : {
            "type" : "string",
            "description" : "Symbolic HTTP response status.",
            "enum" : [ "BAD_REQUEST", "UNAUTHORIZED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "CONFLICT", "PRECONDITION_FAILED", "UNSUPPORTED_MEDIA_TYPE", "UNPROCESSABLE_ENTITY", "PRECONDITION_REQUIRED", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "HTTP_VERSION_NOT_SUPPORTED" ]
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Numeric HTTP response status.",
            "maximum" : 599,
            "minimum" : 400
          },
          "suggestedAction" : {
            "type" : "string",
            "description" : "Recovery action suggested to the client.",
            "enum" : [ "none", "requestRole", "logout", "logoutImmediately", "provideAuthorizationHeader", "refreshToken" ],
            "nullable" : true
          }
        },
        "required" : [ "status", "statusCode" ]
      },
      "CedarGroupReference" : {
        "type" : "object",
        "description" : "A reference to a group. It appears wherever a group is named beside something else, as the holder of a grant or as a group a user belongs to. The name accompanies the identifier so a caller need not fetch the group record to display it.",
        "properties" : {
          "@id" : {
            "type" : "string",
            "format" : "uri",
            "description" : "Identifier of the group."
          },
          "schema:name" : {
            "type" : "string",
            "description" : "Display name of the group."
          }
        },
        "required" : [ "@id" ]
      },
...
      "LogUsageSummary" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "The totals for a range beside the hourly series behind them. The range is echoed back as the\nserver resolved it, so a caller that omitted a bound can see which one it got.\n",
        "properties" : {
          "from" : {
            "type" : "string",
            "description" : "The inclusive lower bound of the range, as a UTC instant."
          },
          "series" : {
            "type" : "array",
            "description" : "One entry per hour that holds a row, oldest first. Hours with no traffic are absent.",
            "items" : {
              "$ref" : "#/components/schemas/TimeBucket"
            }
          },
          "to" : {
            "type" : "string",
            "description" : "The exclusive upper bound of the range, as a UTC instant."
          },
          "totals" : {
            "$ref" : "#/components/schemas/UsageTotals"
          }
        }
      },
...
      "TimeBucket" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "One hour of the request volume series.",
        "properties" : {
          "errorCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "hourUtc" : {
            "type" : "string",
            "description" : "The hour this bucket covers, as a UTC instant on the hour."
          },
          "reqCount" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "UsageTotals" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "Request volume and latency across a whole range. The percentiles are computed from the merged\nlatency histogram of every covered hour, not averaged from per-hour percentiles. Every field\nis zero where no hour in the range holds a row.\n",
        "properties" : {
          "errorCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "p50Nanos" : {
            "type" : "integer",
            "format" : "int64"
          },
          "p95Nanos" : {
            "type" : "integer",
            "format" : "int64"
          },
          "p99Nanos" : {
            "type" : "integer",
            "format" : "int64"
          },
          "reqCount" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "UserDiagnosticReport" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "What each store holds about one user. Empty where the user store does not know the\nidentifier, since the stored profile is what the other sections are gathered from.\n",
        "properties" : {
          "cedarUser" : {
            "type" : "object",
            "additionalProperties" : true,
            "description" : "The stored user profile: identity, home folder, roles, permissions, API keys, UI\npreferences and the authentication source behind the account.\n"
          },
          "keycloak" : {
            "type" : "object",
            "additionalProperties" : false,
            "properties" : {
              "user" : {
                "type" : "object",
                "additionalProperties" : true,
                "description" : "The Keycloak account behind the user, as Keycloak represents it, with its effective\nrealm roles filled in. Null where Keycloak could not be read.\n",
                "nullable" : true
              }
            }
          },
          "neo4j" : {
            "type" : "object",
            "additionalProperties" : false,
            "properties" : {
              "extra" : {
                "type" : "object",
                "additionalProperties" : false,
                "description" : "What the graph answers about the user rather than stores about them.",
                "properties" : {
                  "accessibleCount" : {
                    "allOf" : [ {
                      "$ref" : "#/components/schemas/UserResourceCounts"
                    } ],
                    "description" : "How much of each type the user may reach, counted in the graph."
                  },
                  "groupsWithAdministrator" : {
                    "type" : "array",
                    "items" : {
                      "$ref" : "#/components/schemas/CedarGroupReference"
                    }
                  },
                  "groupsWithMembership" : {
                    "type" : "array",
                    "items" : {
                      "$ref" : "#/components/schemas/CedarGroupReference"
                    }
                  }
                }
              },
              "user" : {
                "allOf" : [ {
                  "$ref" : "#/components/schemas/WorkspaceResourceRecord"
                } ],
                "description" : "The graph's record of the user."
              }
            }
          },
          "opensearch" : {
            "type" : "object",
            "additionalProperties" : false,
            "description" : "How much of each type the search index will return for the user at each role. A count of\n-1 means OpenSearch could not be read; the failure is logged rather than returned. Read\nagainst the graph's own count: the two disagreeing is how a stale permission index shows\nitself.\n",
            "properties" : {
              "editorCount" : {
                "$ref" : "#/components/schemas/UserResourceCounts"
              },
              "managerCount" : {
                "$ref" : "#/components/schemas/UserResourceCounts"
              },
              "viewerCount" : {
                "$ref" : "#/components/schemas/UserResourceCounts"
              }
            }
          },
          "resourceType" : {
            "type" : "string",
            "enum" : [ "user" ]
          }
        }
      },
      "UserResourceCounts" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "One count per artifact type. Template instances are counted under `templateInstance`.",
        "properties" : {
          "element" : {
            "type" : "integer",
            "format" : "int64"
          },
          "field" : {
            "type" : "integer",
            "format" : "int64"
          },
          "folder" : {
            "type" : "integer",
            "format" : "int64"
          },
          "template" : {
            "type" : "integer",
            "format" : "int64"
          },
          "templateInstance" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "WorkspaceResourceRecord" : {
        "type" : "object",
        "additionalProperties" : true,
        "description" : "One node as the workspace graph holds it. The identity, naming and provenance keys below are\ncarried by every kind of node; a folder, an artifact, a group and a user each add their own,\nso the object is left open.\n",
        "properties" : {
          "@context" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "@id" : {
            "type" : "string",
            "format" : "uri"
          },
          "createdOnTS" : {
            "type" : "integer",
            "format" : "int64",
            "nullable" : true
          },
          "lastUpdatedOnTS" : {
            "type" : "integer",
            "format" : "int64",
            "nullable" : true
          },
          "oslc:modifiedBy" : {
            "type" : "string",
            "format" : "uri",
            "nullable" : true
          },
          "pav:createdBy" : {
            "type" : "string",
            "format" : "uri",
            "nullable" : true
          },
          "pav:createdOn" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "pav:lastUpdatedOn" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "resourceType" : {
            "type" : "string"
          },
          "schema:description" : {
            "type" : "string",
            "nullable" : true
          },
          "schema:identifier" : {
            "type" : "string",
            "nullable" : true
          },
          "schema:name" : {
            "type" : "string",
            "nullable" : true
          },
          "sourceHash" : {
            "type" : "string",
            "nullable" : true
          }
        }
      }
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}